Same thing for eigen

+++ include/RigidBody/KalmanReconsIMU.h
@@ -103,7 +103,7 @@ protected:
     virtual void manageOcclusionDuringIteration(
         utils::Matrix &InvTp,
         utils::Vector &measure,
-        const std::vector<unsigned int> &occlusion);
+        const std::vector<size_t> &occlusion);
 
     std::shared_ptr<bool> m_firstIteration; ///< If first iteration was done
 };
